Add "debug" log function
Calling cmdline() needs to run earlier to be useful for logging DETECTED_HOMEDIR and SCRIPTPATH before they are used for logic, so has been moved outside of main(). Could be moved back into main() and just run earlier if needed.
The relocation is untested in a first-run scenario, but did not break for what testing I could do.